tdf#143939 sw: track format changes of the actual word
If there is no text selection, character formattings
are applied on the word under the text cursor,
now with storing the original direct character formatting
in a redline "extra data" for rejection of the tracked
formatting change.

Follow-up to commit ee3171adb24a3d720466fa43bec869910333f892
"tdf#143918 sw: fix reject of multiple changes" (clean-up the
format tracking code into the new lcl_SetRedline(), and
applying that for the not handled case, too).

Change-Id: Ie6bc4dc17dbf84cc61970cef66689aec8ddc4595
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120707
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>