Give gengal invocations their own UserInstallations
Building e.g. Gallery_backgrounds tries to access a UserInstallation at various
places, first at

> #0 0x00007ffff29b0302 in EmbeddedFontsHelper::clearTemporaryFontFiles() () at vcl/source/gdi/embeddedfontshelper.cxx:60
> #1 0x00007ffff2dd6bb4 in InitVCL() () at vcl/source/app/svmain.cxx:304
> #2 0x00007ffff2dd66b6 in ImplSVMain() () at vcl/source/app/svmain.cxx:190
> #3 0x00007ffff2dd6898 in SVMain() () at vcl/source/app/svmain.cxx:231
> #4 0x000000000041215d in sal_main() () at vcl/source/salmain/salmain.cxx:34

It would either access the real UserInstallation (if the bootstrap ini-file has
already been delivered to instdir) or try to access some (typically non-
existent) /user.  That appears to be generally harmless, but to be on the safe
side, pass the gengal invocation an explicit UserInstallation in workdir (which
does not exist, and does not get populated, but just prevents accidentally
accessing a real UserInstallation).

Change-Id: Idc709a4bd44ce0dddbafaac25141215730808fca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120838
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>