tdf#137363 DOCX filter: outlineLvl not associated with iLvl
Lets make sure no one gets the wrong impression that
these are somehow connected together. They aren't at all.

In order to make that work, the newly added para style
ability to carry PROP_NUMBERING_LEVEL should not be trusted
as an inheritance, since it can drift out of sync when
ValidateListlevel changes the value to body level.

Lets just ignore the drift for now.

Unfortunately, I don't see any way in UNO to setPropertyToDefault
or Erase or ClearProperty.

Keeping PROP_NUMBERING_LEVEL in sync with 9 isn't great
because 9 has a totally different meaning in LO UI.
There it means the highest level 10. This can be seen when enabling
numbering on these paragraphs - they jump to level 10 right away
since they already have the property set to 9, the highest value.

The other option would be setting to 0 (the default value).
But that still doesn't solve the out-of-sync/inheritance problem.

Thus I just ignore the problem for now, waiting for
-PROP_NUMBERING_LEVEL to gain a NO_LEVEL value
-uno gaining an erase ability
-a real problem that forces some kind of solution.

Change-Id: Ie6d098fa60a18ac2e3e7d1e33276ca923bfbfaa5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121368
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>