[API CHANGE] Fix all bad UNOIDL identifiers across offapi
Identifiers containing underscores must start with an uppercase letter.  (So
that the UNO implementation can use identifiers containing underscores and
starting with a lowercase letter for internal purposes, see e.g. the static_type
member functions in the C++ *.hdl files.  idlc checks that with the -cid option,
while unoidl-write silently allows bad identifiers for now, see the TODO in
unoidl/source/sourceprovider-scanner.l, which can be dropped after this change.)

All of the affected identifiers were present since early OOo times, but none of
them appear to be used much, at least not across LO itself, so there is hope
that we can get away with these incompatible changes.  (For the constant group
members, we could roll this out in two steps, introducing the new members
alongside the old ones and deprecating the old ones in a first step, then
removing the old, deprecated ones in a second step, so that third-party code
would have more time to adapt.  But that would not work for the enum members and
interface methods, so just do all of this in one breaking step.)

udkapi happened to not contain any bad identifiers.

Change-Id: If2d4c16563606f9efb48b937c76af54746377428
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121725
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>