wina11y: Move "stdafx.h" include to headers
Move the "stdafx.h" include to the MAccessible.h
and UNOXWrapper.h headers from the corresponding
source files, since the 'ATL_NO_VTABLE' define
from stdafx.h is already used in the headers.

(This e.g. also makes sure that the define is known
when opening one of those headers in an IDE.)

Change-Id: I3b6b7df037846daaad4d702e0e0c0bd2970dc663
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121773
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>