add a oss-fuzz-setup.sh to move setting up oss-fuzz env into our git
instead of making an upstream change for each dependency change.

We already download missing dependencies within our build, but in
oss-fuzz we have multiple builds back to back in fresh environments, so
we would do that for each build, e.g. once for asan, again for ubsan and
again for msan etc.

Change-Id: Ie810df5e84d4b265200e5c54831a4da43255af55
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122598
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>