tdf#144317 sw table minimize: fix signed->unsigned table growth
In the case where the table was already rather oversized,
the minimize and optimize functions would attempt to shrink
the table to the maximum recommended size. But that was done
without checking whether that would result in a negative
position for the column boundary. (At least that is what I
think was happening. This code is a bit too cryptic...)

I tried to make a unit test, but the table size didn't
grow or shrink as I expected it to (and as it seems to
when the command is run by hand...)
An F12 with SW_DEBUG=true ./instdir/program/soffice
shows that the cell gains a huge width, but that didn't
show up either in a parseDump, even after a
Scheduler::ProcessEventsToIdle(). So I gave up (again).

Change-Id: Id4b9defae718694aa76a3db01f6b02ead5f98e6b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123108
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>