don't warn about missing TopMostParentSystemWindow
Seen in tdf#144037 because the notebookbar is not a child of the
toplevel window, but instead (like the menubar) is a child of
borderwindow of the toplevel so from its perspective there is no
eventual parent SystemWindow.

This TaskPane thing could probably be dropped and likely no one would
notice. But with tearable floaters like the color dropdown still a thing
then F6 wouldn't work to move into them and someone might complain.

Change-Id: I7174598a2ee01e98aca2f6e214adf94f7a555018
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123258
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>