tdf#136715 sw: fix table Undo to update table style
Applying table styles works by applying direct formatting
on the table. Some of the built-in table styles, like
"Box List Red" specifies boldness for the last row of
the table. After selecting and deleting the last n rows
of a table, which has a table style which specifies formatting
for the last row, the new last row gets these formatting as
expected. However before the fix, after undoing the deletion
of the rows, this formatting were still applied on that row
which previously became the last row.

Change-Id: I3ad086dd9b5e4111b3f019ca3d23f1db8b348637
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123730
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>