sw: avoid popup window for the EmbeddedObject property of an UNO OLE object
The old behavior was that in case loading the native data of the
embedded object failed, then we presented a popup to the user, but we
still returned a valid com.sun.star.comp.embed.OCommonEmbeddedObject.

Given that the return valid is desired even in the error case, don't
turn the error popup into an UNO exception, rather redirect the popup to
SAL_WARN().

This is meant to be consistent with other UNO API functions that pass
down an "is API" boolean to internal code to control if interactive
behavior is allowed or not.

One extreme would be to do this only for the EmbeddedObject property of
frames, the other one would be to do this for more object types. Go with
a middle ground for now, so all properties of Writer frame objects use
this, but leave other object types unchanged.

Change-Id: I655fcb3780e96cecc7ed3f01ded52948d013172e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/123752
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins